myproject/trainee/requirements.txt
Django==5.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.27.0
django-filter==24.1
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
pymongo==4.6.1
boto3==1.34.61
botocore==1.34.61
python-pptx==0.6.23
reportlab==4.0.9